Add AESCipher in different file
